+++ packages/client-ts/src/context/utils.ts
@@ -44,7 +44,9 @@ export class RecordSignature {
         }
     
         const sig = await keyring.sign(_data)
-        if (!data.signatures[signKey.toLowerCase()]) {
+
+        // Create empty signature object if this DID hasn't signed, or if this DID has an old signature format (string, not object)
+        if (!data.signatures[signKey.toLowerCase()] || typeof(data.signatures[signKey.toLowerCase()]) === 'string') {
           data.signatures[signKey.toLowerCase()] = {}
         }