+++ packages/vda-common/src/defaults.ts
@@ -4,11 +4,15 @@ export function getWeb3ConfigDefaults(chainName: string) {
     switch (chainName) {
         case 'devnet':
             return {
-                rpcUrl: RPC_URLS[chainName]
+                rpcUrl: RPC_URLS[chainName],
+                eip1559Mode: 'fast',
+                eip1559gasStationUrl: 'https://gasstation-testnet.polygon.technology/v2'
             }
         case 'testnet':
             return {
-                rpcUrl: RPC_URLS[chainName]
+                rpcUrl: RPC_URLS[chainName],
+                eip1559Mode: 'fast',
+                eip1559gasStationUrl: 'https://gasstation-testnet.polygon.technology/v2'
             }
         case 'mainnet':
             return {